+++ docs/lib-entry.ts
@@ -2,5 +2,6 @@
 // from dropping re-exports of a package marked "sideEffects": false.
 import * as helpers from '../src/helpers/index.ts';
 import * as edgewise from '../src/index.ts';
+import * as worker from '../src/worker/index.ts';
 
-export default { ...edgewise, helpers };
+export default { ...edgewise, helpers, worker };